refactor: rework after review
